remove predict_postings.py
